Make X509CRLStreamWriter an interface.
This is in preparation for adding a JSS implementation.
